Moving run_custom_function command to its own test file
